Moved InvalidTelephoneNumberException out of SMSPeer, it wasn't public so couldn't be caught properly
